// src/components/PreferencesAndEnvironment/PreferencesAndEnvironment.types.ts
export type PreferencesAndEnvironment = {
  wallDecor: string;
  preferredFurniture: string;
  dominantColors: string;
  regularActivities: string[];
  adviceSource: string;
  transportAudio: string;
  preferredChannels: string;
  aloneTimeActivities: string;
  preferredSports: string;
  socialMediaFrequency: string;
  socialMediaDetox: string;
};